src/streamlit/requirements.txt
pyarrow==14.0.0
numpy
streamlit==1.36.0
altair==4
pillow